% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllPlots.R
\name{AllPlots,folder-method}
\alias{AllPlots,folder-method}
\title{All plot folder}
\usage{
\S4method{AllPlots}{folder}(
  object,
  pdf = T,
  pdfall = F,
  covPrim = "none",
  covSec = "none",
  runDataDes = 0,
  addFunction = "",
  sensitivity = ""
)
}
\description{
All plot folder
}
