% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllPlots.R
\name{AllPlots,run-method}
\alias{AllPlots,run-method}
\title{All plot single run}
\usage{
\S4method{AllPlots}{run}(
  object,
  pdf = T,
  pdfall = F,
  covPrim = "none",
  covSec = "none",
  addFunction = ""
)
}
\description{
All plot single run
}
