% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllPlots.R
\name{AllPlots}
\alias{AllPlots}
\title{Output: a pdf with lots of plots}
\usage{
AllPlots(
  object,
  pdf = T,
  pdfall = F,
  covPrim = "none",
  covSec = "none",
  runDataDes = 0,
  addFunction = "",
  sensitivity = ""
)
}
\description{
Output: a pdf with lots of plots
}
