% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_forest.R
\name{plot_forest_run}
\alias{plot_forest_run}
\title{Plot forest_run}
\usage{
plot_forest_run(object, n = 1000, returnExpr = F, onlytable = F)
}
\value{

}
\description{
Plot forest_run
}
\author{
Thibaud Derippe
}
