% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3SetReplace_ADAPT.R
\name{setPath_ADAPT<-,run-method}
\alias{setPath_ADAPT<-,run-method}
\title{SetReplaceMethod to fill a run object with the corresponding patwhay
Author: Thibaud Derippe
Arguments: object, the run object we want to fill the attribute
Arguments: value, pathway to the run
Output: not concerned, SetReplacedMethod}
\usage{
\S4method{setPath_ADAPT}{run}(object) <- value
}
\description{
SetReplaceMethod to fill a run object with the corresponding patwhay
Author: Thibaud Derippe
Arguments: object, the run object we want to fill the attribute
Arguments: value, pathway to the run
Output: not concerned, SetReplacedMethod
}
