% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3SetReplace_NLMIXR_Run.R
\name{setPath_NLMIXR<-,run-method}
\alias{setPath_NLMIXR<-,run-method}
\title{Author: Thibaud Derippe
Arguments: object, the run object we want to fill the attribute
Arguments: value, pathway to the run
Output: not concerned, SetReplacedMethod}
\usage{
\S4method{setPath_NLMIXR}{run}(object) <- value
}
\description{
Author: Thibaud Derippe
Arguments: object, the run object we want to fill the attribute
Arguments: value, pathway to the run
Output: not concerned, SetReplacedMethod
}
